refactor, move to meta_master.proto
