+++ core/transport/src/main/proto/grpc/meta_master.proto
@@ -190,25 +190,27 @@ message GetConfigHashPResponse {
   * to query cluster configuration.
   */
 service MetaMasterConfigurationService {
-    /**
-     * Returns a list of Alluxio runtime configuration information.
-     */
-    rpc GetConfiguration (GetConfigurationPOptions) returns (GetConfigurationPResponse);
+  /**
+   * Returns a list of Alluxio runtime configuration information.
+   */
+  rpc GetConfiguration (GetConfigurationPOptions) returns (GetConfigurationPResponse);
 
-    /**
-     * Sets a property for a path.
-     */
-    rpc SetPathConfiguration (SetPathConfigurationPRequest) returns (SetPathConfigurationPResponse);
+  /**
+   * Sets a property for a path.
+   */
+  rpc SetPathConfiguration (SetPathConfigurationPRequest) returns (SetPathConfigurationPResponse);
 
-    /**
-     * Removes properties for a path, if the keys are empty, it means remove all properties.
-     */
-    rpc RemovePathConfiguration (RemovePathConfigurationPRequest) returns (RemovePathConfigurationPResponse);
+  /**
+   * Removes properties for a path, if the keys are empty, it means remove all properties.
+   */
+  rpc RemovePathConfiguration (RemovePathConfigurationPRequest) returns (RemovePathConfigurationPResponse);
 
-    /**
-     * Returns the hashes of cluster and path level configurations.
-     */
-    rpc GetConfigHash (GetConfigHashPOptions) returns (GetConfigHashPResponse);
+  /**
+   * Returns the hashes of cluster and path level configurations.
+   */
+  rpc GetConfigHash (GetConfigHashPOptions) returns (GetConfigHashPResponse);
+
+  rpc UpdateConfig(UpdateConfigPRequest) returns (UpdateConfigPResponse);
 }
 
 message GetMasterIdPOptions {}
@@ -269,6 +271,4 @@ service MetaMasterMasterService {
    * Heartbeats to indicate the master is lost or not.
    */
   rpc MasterHeartbeat(MasterHeartbeatPRequest) returns (MasterHeartbeatPResponse);
-
-  rpc UpdateConfig(UpdateConfigPRequest) returns (UpdateConfigPResponse);
 }
